requirements.txt
python-dotenv>=0.21.0
requests>=2.31.0
pytest>=7.0.0
pandas>=1.5.0
reportlab>=3.6.0
fpdf>=1.7.2
Pillow>=9.0.0
PyPDF2>=3.0.0
docx2pdf>=0.1.8